+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.15.0
 ### Features
 * add new API event touchscroll for mobile (use it in less) [#556](https://github.com/jcubic/jquery.terminal/issues/556)
+* mobile paste [#458](https://github.com/jcubic/jquery.terminal/issues/458)
 ### Bugfix
 * fix vertical bar cursor animation on empty command line
 * fix edge case while splitting the command line with formatting (better fix for [#379](https://github.com/jcubic/jquery.terminal/379))